// src/infrastructure/Enums/EarthquakePolicyType.ts
enum EarthquakePolicyType {
    SQUARE_METERS = 'SQUARE_METERS',
    TOTAL_FLOORS = 'TOTAL_FLOORS', 
    CONSTRUCTION_YEAR = 'CONSTRUCTION_YEAR',
    CONSTRUCTION_STYLE = 'CONSTRUCTION_STYLE',
    EARTHQUAKERISK = 'EARTHQUAKERISK',
    CONSTANT = 'CONSTANT',
    POLICY_TYPE = 'POLICY_TYPE'
}

export default EarthquakePolicyType;